+++ pytsite/core/router.py
@@ -152,7 +152,7 @@ def dispatch(env: dict, start_response: callable):
         rule, rule_args = __url_adapter.match(return_rule=True)
 
         # Notify listeners
-        events.fire('router.dispatch')
+        events.fire('pytsite.core.router.dispatch')
 
         # Processing rule filters
         for flt in rule.filters: