+++ pytsite/settings/_functions.py
@@ -11,19 +11,23 @@ __settings = {}
 
 
 def define(uid: str, form_cls: type, menu_title: str, menu_icon: str, menu_weight: int=0):
+    """Define setting.
+    """
     if uid in __settings:
         raise KeyError("Setting '{}' already defined.".format(uid))
 
     if not isinstance(form_cls, type) or not issubclass(form_cls, _form.Base):
         raise TypeError("Subclass of base form expected.")
 
-    __settings[uid] = {'title': menu_title,  'form_cls': form_cls, 'weight': menu_weight}
+    __settings[uid] = {'title': menu_title, 'form_cls': form_cls, 'weight': menu_weight}
 
     url = _router.endpoint_url('pytsite.settings.eps.form', {'uid': uid})
-    admin.sidebar.add_menu('settings', uid, menu_title, url, menu_icon, permissions=('*',))
+    admin.sidebar.add_menu('settings', uid, menu_title, url, menu_icon, permissions='*')
 
 
-def get_definition(uid: str):
+def get_definition(uid: str) -> dict:
+    """Get setting definition.
+    """
     if uid not in __settings:
         raise KeyError("Setting '{}' is not defined.".format(uid))
 
@@ -31,6 +35,8 @@ def get_definition(uid: str):
 
 
 def get_form(uid) -> _form.Base:
+    """Get form for setting.
+    """
     frm_class = get_definition(uid)['form_cls']
     frm = frm_class('settings-' + uid)
     """:type : _form.Base """
@@ -62,6 +68,8 @@ def get_form(uid) -> _form.Base:
 
 
 def get_setting(uid) -> dict:
+    """Get setting value.
+    """
     if uid not in __settings:
         raise KeyError("Setting '{}' is not defined.".format(uid))
 
@@ -73,6 +81,8 @@ def get_setting(uid) -> dict:
 
 
 def set_setting(uid, value: dict):
+    """Set setting value.
+    """
     entity = _odm.find('setting').where('uid', '=', uid).first()
     if not entity:
         entity = _odm.dispense('setting').f_set('uid', uid)